(PC-26883)[API] refactor: remove unused users from industrial sandbox
